Backend/requirements.txt
django
requests
PyJWT
django-cors-headers
python-dotenv